Enable in all prog-mode buffers by default
Previously the global mode only caused the local mode to be activated
in `emacs-lisp-mode' buffers.  That is a pretty bad default, which I
picked a long time ago when I was the only user and needed just that.

Even though it is this package's only option, many users did not find
`hl-todo-activate-in-modes', making a less surprising default highly
desirable.