No longer treat incomplete searches as non-matches
We search for text that matches the keyword regexp.  When there is
a match, then we checked whether the corresponding text is within
a comment or doc-string.  If so, then we considered that a match.

Otherwise, the regexp search succeeded but the syntax check failed,
then we used to consider that a non-match, which is incorrect; what
we have in this case is an incomplete search, not a failed search.

In that case we have to finish the search from the current position.
This can happen repeatedly, so we implement it using a new function
`hl-todo--search', which calls itself to resume an incomplete search.

Fixes #26.