Only enable mode in buffers that use font-lock-mode
`hl-todo-mode' depends on `font-lock-mode'.  Enabling the former in
a buffer that does not use the latter likely removes text properties
that were put in place using some other mechanism.  So now we abort
if this is about to happen.

We cannot just check the value of `font-lock-mode' because apparently
that may be non-nil even though that mode is not actually in effect
or about to be put into effect.  It is unclear why that is so.  It
is also unclear why we already did check for `font-lock-mode' (but
seemingly too late).

At least `font-lock-specified-p' seems to serve our purpose, so use
that.

See #55 and https://debbugs.gnu.org/cgi/bugreport.cgi?bug=42138.