hl-todo-mode: Speed up
Previously, in an effort to speed up the process, we didn't
refontify the whole buffer.  Instead we only added our additional
fontification. However, we did not try to limit the latter to parts
that were already fontified.  Instead we searched the whole buffer,
and that is costly too.

Now we discard all existing fontification and then immediately
refontify the visible portion using all keywords, including but not
limited to, the ones that we just added.  Other parts of the buffer
are refontified lazily.  Re-enabling `jit-lock-mode' does all that.

Closes #61.