No longer link to alternative implementations
The idea behind that link was to redirect users who want more features
to packages that provide more features, but that has never actually
worked.  This package seems to be the de facto default for this sort
of functionality and if something is missing, someone sooner or later
requests that it is added here.  So remove the passive-aggressive and
ineffective comment and link.