hl-todo--turn-on-mode-if-desired: Not in temporary buffers
This prevents it from being turned on in the temporary buffer used by
`magit-revision--wash-message'.  If we highlight keywords this early
and by using `hl-todo-mode', then it happens before the heading of the
section is highlighting, which then fails to set the background for
any part in the heading that is a keyword.  Keywords are supposed to
be highlighted later, using `hl-todo-search-and-highlight'.