hl-todo-search-and-highlight: Replace hl-todo-magit-revision
Highlight regardless of whether `global-hl-todo-mode' is enabled or
not.  This makes the function useful in other contexts, so rename it
accordingly.  Do not add to magit hook by default; instead encourage
users to do so.  In summary; make it more flexible, at the cost of
requiring the user to opt-in.